+++ figure_parser/core/parser/base.py
@@ -1,18 +1,18 @@
-from abc import abstractmethod
-from typing import List, Optional, Protocol, Type, TypeVar
+from abc import ABC, abstractmethod
+from typing import Generic, List, Optional, Type, TypeVar
 
 from figure_parser.core.entity import OrderPeriod, Release
 
-Source_T = TypeVar('Source_T', contravariant=True)
+Source_T = TypeVar('Source_T')
 Parser_T = TypeVar('Parser_T')
 
 
 __all__ = (
-    'ProductParserInterface',
+    'AbstractProductParser',
 )
 
 
-class ProductParserInterface(Protocol[Source_T]):
+class AbstractProductParser(ABC, Generic[Source_T]):
     """Abstract product parser class"""
     @classmethod
     @abstractmethod
