+++ figure_parser/pipes.py
@@ -1,0 +1,3 @@
+from figure_parser.core.pipe.nomalization_pipe import (
+    ProductGeneralFieldstNormalizer, ProductWorkerFieldstNormalizer)
+from figure_parser.core.pipe.sorting_pipe import ReleaseSorting
